build(deps): update pdfium-render, scraper, sqlx and mockall to new majors
- Follow `pdfium-render` 0.9: page indices are `i32`, and `PdfBitmap::as_image`
  now returns a `Result`
- Replace sqlx's removed `runtime-tokio-native-tls` feature with `runtime-tokio`,
  dropping the TLS stack a SQLite-only build never used
- Mark the returned `id` of the bookmark insert as non-null, because sqlx 0.9 no
  longer takes a rowid alias's `NOT NULL` from the schema
- Regenerate `.sqlx` for the 0.9 offline query format

+++ src-tauri/src/perfbench.rs
@@ -249,7 +249,7 @@ fn build_pdf(dir: &Path, pages: &[Vec<u8>]) -> PathBuf {
         obj.scale(w.value, h.value).expect("scale");
         let mut page = doc
             .pages_mut()
-            .create_page_at_index(PdfPagePaperSize::a4(), i as u16)
+            .create_page_at_index(PdfPagePaperSize::a4(), i as i32)
             .expect("create page");
         page.objects_mut().add_image_object(obj).expect("add image");
     }
